Renamed catalog source to kedro_data_catalog
