+++ kedro/io/__init__.py
@@ -15,7 +15,7 @@ from .core import (
     Version,
 )
 from .data_catalog import DataCatalog
-from .data_catalog_redesign import BaseDataCatalog, KedroDataCatalog
+from .kedro_data_catalog import BaseDataCatalog, KedroDataCatalog
 from .lambda_dataset import LambdaDataset
 from .memory_dataset import MemoryDataset
 from .shared_memory_dataset import SharedMemoryDataset
